delated notes controller file
